linux: uncontroversial cleanups found by clang
Part of a larger patch to use clang to warning-check our code.

Review URL: http://codereview.chromium.org/561008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37781 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f884f2a323456f31486f9dd33d704ee086475355

+++ base/sys_string_conversions_linux.cc
@@ -40,7 +40,7 @@ std::string SysWideToNativeMB(const std::wstring& wide) {
     size_t res = src ? wcrtomb(buf, src, &ps) : 0;
     switch (res) {
       // Handle any errors and return an empty string.
-      case -1:
+      case static_cast<size_t>(-1):
         return std::string();
         break;
       case 0:
@@ -68,7 +68,7 @@ std::string SysWideToNativeMB(const std::wstring& wide) {
     size_t res = src ? wcrtomb(&out[j], src, &ps) : 0;
     switch (res) {
       // Handle any errors and return an empty string.
-      case -1:
+      case static_cast<size_t>(-1):
         return std::string();
         break;
       case 0:
@@ -96,8 +96,8 @@ std::wstring SysNativeMBToWide(const StringPiece& native_mb) {
     size_t res = mbrtowc(NULL, src, native_mb.size() - i, &ps);
     switch (res) {
       // Handle any errors and return an empty string.
-      case -2:
-      case -1:
+      case static_cast<size_t>(-2):
+      case static_cast<size_t>(-1):
         return std::wstring();
         break;
       case 0:
@@ -125,8 +125,8 @@ std::wstring SysNativeMBToWide(const StringPiece& native_mb) {
     size_t res = mbrtowc(dst, src, native_mb.size() - i, &ps);
     switch (res) {
       // Handle any errors and return an empty string.
-      case -2:
-      case -1:
+      case static_cast<size_t>(-2):
+      case static_cast<size_t>(-1):
         return std::wstring();
         break;
       case 0:
